Consolidate more tests
